telegram/first/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
pyaes==1.6.1
    # via telethon
pyasn1==0.6.0
    # via rsa
python-dotenv==1.0.1
    # via -r requirements.in
rsa==4.9
    # via telethon
telethon==1.36.0
    # via -r requirements.in
